+++ gemrb/plugins/Core/Game.h
@@ -15,7 +15,7 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA 02111-1307, USA.
  *
- * $Header: /data/gemrb/cvs2svn/gemrb/gemrb/gemrb/plugins/Core/Game.h,v 1.59 2005/06/26 13:57:51 avenger_teambg Exp $
+ * $Header: /data/gemrb/cvs2svn/gemrb/gemrb/gemrb/plugins/Core/Game.h,v 1.60 2005/06/28 20:08:13 guidoj Exp $
  *
  */
 
@@ -167,7 +167,7 @@ public:
 	/* returns the npcs count */
 	int GetNPCCount() const { return (int)NPCs.size(); }
 	/* sends the hotkey trigger to all selected pcs */
-	void SetHotKey(int Key);
+	void SetHotKey(unsigned long Key);
 	/* select PC for non-walking environment (shops, inventory, ...) */
 	bool SelectPCSingle(int index);
 	/* get index of selected PC for non-walking env (shops, inventory, ...) */
